Added comments to rng_sample.md about modulo bias
Added comments about the following:
max_value is wrong
the code suffers heavily from modulo bias
the article does not mention modulo bias
====
The WAX RNG Tutorial rng_sample gives an example of getting a random number (0-99) using a byte. But it doesn't mention modulo bias anywhere. `uint8_t random_int % 100` gives a +50% bias to the first 56 numbers (0-55); a byte isn't large enough to mitigate modulo bias in this case. Modulo bias is an important subject when it comes to RNG, it can be eliminated with rejections, proper modulo boundaries or mitigated with large storage space vs. modulus size (e.g. if random_int's data type were uint64_t).

Using a single byte for this would be a bad practice to follow and teach. And I think it might be good for the tutorial to mention modulo biases.

Also it says the max_value is 100 but the max value is probably actually 99. (the largest X % 100 can be is 99).

And the Twitter link in the site's footer leads to the discord, not twitter.